Add files via upload
This  is an example  for  the  Message   encoding and  decoding  using  cryptography.fernet 

Fernet guarantees that a message encrypted using it cannot be manipulated or read without the key. Fernet is an implementation of symmetric (also known as “secret key”) authenticated cryptography. Fernet also has support for implementing key rotation via MultiFernet.